+++ examples/spot/funding_account/DepositAddress/DepositAddress.go
@@ -6,8 +6,8 @@ import (
 )
 
 /*
-	GET /account/v1/deposit/address
-	Doc: https://developer-pro.bitmart.com/en/spot/#deposit-address-keyed
+GET /account/v1/deposit/address
+Doc: https://developer-pro.bitmart.com/en/spot/#deposit-address-keyed
 */
 func main() {
 
@@ -23,7 +23,7 @@ func main() {
 	if err != nil {
 		log.Panic(err)
 	} else {
-		log.Println(bitmart.GetResponse(ac))
+		log.Println(ac.Response)
 	}
 
 }