+++ examples/spot/funding_account/MarginAccountDetails/MarginAccountDetails.go
@@ -6,8 +6,8 @@ import (
 )
 
 /*
-	GET /spot/v1/margin/isolated/account
-	Doc: https://developer-pro.bitmart.com/en/spot/#get-margin-account-details-isolated-keyed
+GET /spot/v1/margin/isolated/account
+Doc: https://developer-pro.bitmart.com/en/spot/#get-margin-account-details-isolated-keyed
 */
 func main() {
 
@@ -19,12 +19,19 @@ func main() {
 	})
 
 	// Get Margin Account Details(Isolated) (KEYED)
-	var ac, err = client.GetMarginAccountDetailsIsolated("BTC_USDT")
-
+	var ac, err = client.GetMarginAccountDetailsIsolated()
 	if err != nil {
 		log.Panic(err)
 	} else {
-		log.Println(bitmart.GetResponse(ac))
+		log.Println(ac.Response)
 	}
 
+	var ac2, err2 = client.GetMarginAccountDetailsIsolated(map[string]interface{}{
+		"symbol": "BTC_USDT",
+	})
+	if err2 != nil {
+		log.Panic(err2)
+	} else {
+		log.Println(ac2.Response)
+	}
 }